Quản Lý Nhà Hàng - Winfrom_C#

+++ RestaurantManagement.BUS/CustomerService.cs
@@ -1,35 +1,40 @@
 ﻿using RestaurantManagement.DAL.Entities;
 using System.Collections.Generic;
+using System.Data.Entity.Migrations;
 using System.Linq;
 
 namespace RestaurantManagement.BUS
 {
     public class CustomerService
     {
+        private readonly RestaurantManagementDBContext context = new RestaurantManagementDBContext();
         public List<Customer> GetAll()
         {
-            RestaurantManagementDBContext context = new RestaurantManagementDBContext();
             return context.Customers.ToList();
-        }
-
-        public Customer FindById(string id)
-        {
-            RestaurantManagementDBContext context = new RestaurantManagementDBContext();
-            return context.Customers.FirstOrDefault(p => p.MaKH == id);
         }
 
         public void InsertUpdate(Customer customer)
         {
-            RestaurantManagementDBContext context = new RestaurantManagementDBContext();
-            context.Customers.Add(customer);
+            context.Customers.AddOrUpdate(customer);
             context.SaveChanges();
         }
 
-        public void DeleteUpdate(Customer customer)
+        public void Delete(string makh)
         {
-            RestaurantManagementDBContext context = new RestaurantManagementDBContext();
-            context.Customers.Remove(customer);
-            context.SaveChanges();
+            var customer = context.Customers.FirstOrDefault(s => s.MaKH == makh);
+            if (customer != null)
+            {
+                context.Customers.Remove(customer);
+                context.SaveChanges();
+            }
+        }
+        public Customer FindByMaKH(string makh)
+        {
+            return context.Customers.FirstOrDefault(c=>c.MaKH == makh);
+        }
+        public Customer FindByName(string name)
+        {
+            return context.Customers.FirstOrDefault(c=>c.TenKH.Contains(name));
         }
     }
 }
